BA-1832: fixes
