+++ packages/components/modules/profiles/ProfilePopover/AddProfileMenuItem/types.ts
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { AddProfileModalProps } from '../AddProfileModal'
+import { AddProfileModalProps } from '../AddProfileModal/types'
 
 export interface AddProfileMenuItemProps {
   addNewProfileLabel?: string
