+++ packages/components/modules/profiles/ProfilePopover/AddProfileModal/index.tsx
@@ -1,57 +1,31 @@
 import { FC } from 'react'
 
-import { TextField } from '@baseapp-frontend/design-system'
+import { CloseIcon, TextField } from '@baseapp-frontend/design-system'
 import { setFormRelayErrors } from '@baseapp-frontend/utils'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 import { LoadingButton } from '@mui/lab'
 import {
+  Box,
   Button,
   Dialog,
   DialogActions,
   DialogContent,
   DialogTitle,
+  IconButton,
   Link,
   Typography,
 } from '@mui/material'
 import { useForm } from 'react-hook-form'
 import { ConnectionHandler } from 'react-relay'
 import slugify from 'slugify'
-import z from 'zod'
 
-import { Form } from '../../../__shared__/SocialInput/styled'
 import { useOrganizationCreateMutation } from '../../graphql/mutations/OrganizationCreate'
-
-const slugRegex = /^[a-z0-9]+(?:-[a-z0-9]+)*$/
-
-const schema = z.object({
-  name: z
-    .string()
-    .trim()
-    .min(3, { message: 'Must have at least three character' })
-    .max(255, { message: 'Must have at most 255 characters' }),
-  urlPath: z
-    .string()
-    .regex(slugRegex, 'Must contain only lowercase letters, numbers, and hyphens')
-    .max(500, { message: 'Must have at most 500 characters' }),
-})
+import { schema } from './constants'
+import { Form } from './styled'
+import { AddProfileModalProps, OrganizationCreateForm } from './types'
 
 let nextClientMutationId = 0
-
-export interface AddProfileModalProps {
-  addNewProfileLabel?: string
-  termsAndConditionsUrl?: string
-  addNewProfileDescription?: string
-  onClose?: () => void
-  open: boolean
-  setOpen: (open: boolean) => void
-  submitLabel?: string
-}
-
-interface OrganizationCreateForm {
-  name: string
-  urlPath: string
-}
 
 const AddProfileModal: FC<AddProfileModalProps> = ({
   addNewProfileLabel = 'New organization',
@@ -97,9 +71,7 @@ const AddProfileModal: FC<AddProfileModalProps> = ({
         if (!payload) {
           return
         }
-
         const newEdge = payload.getLinkedRecord('profile')
-
         const root = store.getRoot()
         const meRecord = root?.getLinkedRecord('me')
 
@@ -153,41 +125,54 @@ const AddProfileModal: FC<AddProfileModalProps> = ({
         },
       }}
     >
-      <DialogTitle id="organization-modal-title">{addNewProfileLabel}</DialogTitle>
-      <Form onSubmit={form.handleSubmit(onSubmit)} sx={{ border: 'none', margin: 0 }}>
+      <DialogTitle id="organization-modal-title">
+        <Box
+          sx={{ display: 'flex', alignItems: 'center', gap: 2, justifyContent: 'space-between' }}
+        >
+          <Typography variant="h6">{addNewProfileLabel}</Typography>
+          <IconButton onClick={handleClose}>
+            <CloseIcon />
+          </IconButton>
+        </Box>
+      </DialogTitle>
+      <Form onSubmit={form.handleSubmit(onSubmit)}>
         <DialogContent
           sx={{ display: 'flex', flexDirection: 'column', gap: 4 }}
           id="organization-modal-description"
         >
           <Typography color="text.secondary">{addNewProfileDescription}</Typography>
-          <TextField
-            label="Name"
-            name="name"
-            control={form.control}
-            onChange={() => {
-              form.trigger('name')
-            }}
-            onBlur={() => {
-              if (!form.getValues('urlPath')) {
-                form.setValue('urlPath', slugify(form.getValues('name').toLocaleLowerCase()))
+          <Box display="flex" flexDirection="column" gap={2}>
+            <TextField
+              label="Name"
+              name="name"
+              size="medium"
+              control={form.control}
+              onChange={() => {
+                form.trigger('name')
+              }}
+              onBlur={() => {
+                if (!form.getValues('urlPath')) {
+                  form.setValue('urlPath', slugify(form.getValues('name').toLocaleLowerCase()))
+                  form.trigger('urlPath')
+                }
+              }}
+            />
+            <TextField
+              label="URL Path"
+              name="urlPath"
+              size="medium"
+              control={form.control}
+              onChange={() => {
                 form.trigger('urlPath')
-              }
-            }}
-          />
-          <TextField
-            label="Url Path"
-            name="urlPath"
-            control={form.control}
-            onChange={() => {
-              form.trigger('urlPath')
-            }}
-          />
-          <Typography color="text.primary">
-            Upon confirming, you agree to our{' '}
-            <Link display="inline" href={termsAndConditionsUrl} target="_blank">
-              Terms and Conditions.
-            </Link>
-          </Typography>
+              }}
+            />
+            <Typography color="text.primary">
+              Upon confirming, you agree to our{' '}
+              <Link display="inline" href={termsAndConditionsUrl} target="_blank">
+                Terms and Conditions.
+              </Link>
+            </Typography>
+          </Box>
         </DialogContent>
         <DialogActions sx={{ justifyContent: 'flex-end' }}>
           <Button variant="outlined" color="inherit" onClick={handleClose} sx={{ width: 'auto' }}>
