+++ packages/components/modules/profiles/ProfilePopover/AddProfileModal/types.ts
@@ -1,0 +1,14 @@
+export interface AddProfileModalProps {
+  addNewProfileLabel?: string
+  termsAndConditionsUrl?: string
+  addNewProfileDescription?: string
+  onClose?: () => void
+  open: boolean
+  setOpen: (open: boolean) => void
+  submitLabel?: string
+}
+
+export interface OrganizationCreateForm {
+  name: string
+  urlPath: string
+}
